requirements.txt
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dgl-cu113==0.8.1
dglgo==0.0.1
dill==0.3.4
docutils==0.17.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
ntlm-auth==1.5.0
numpy==1.22.3
numpydoc==1.2.1
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
Pygments==2.11.2
pylatexenc==2.10
pynvim==0.4.3
pyparsing==3.0.8
pyrsistent==0.18.1
python-constraint==1.4.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qiskit==0.36.1
qiskit-aer==0.10.4
qiskit-ibmq-provider==0.19.1
qiskit-ignis==0.7.0
qiskit-terra==0.20.1
requests==2.27.1
requests-ntlm==1.1.0
retworkx==0.11.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
stevedore==3.5.0
symengine==0.9.2
sympy==1.10.1
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
tweedledum==1.1.1
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.3
widgetsnbextension==3.6.0
zipp==3.8.0
